Introduce sortpom-maven-plugin.
This ensure that the pom.xml is sorted and formatted according to
the standard. This makes it slightly easier to navigate the pom
as it conforms to expectation and reduce unnecessary ordering
changes so that diffs are easier to parse.

Note that the plugin only veryfies the order by default. To sort
the pom according tot he standard, you can use `./mvnw sortpom:sort`.